pyproject.toml
[tool.poetry]
name = "sports-quant"
version = "1.1.1"
description = "NFL data pipeline combining PFF grades and PFR game data for over/under analysis"
authors = ["Thad Hutcheson <thadhutch@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/thadhutch/sports-quant"
repository = "https://github.com/thadhutch/sports-quant"
keywords = ["nfl", "football", "data-pipeline", "pff", "sports-analytics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "sports_quant", from = "src"}]

[tool.poetry.scripts]
sports-quant = "sports_quant.cli:cli"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.23.1"
pandas = ">=2.2.2,<4.0.0"
python-dotenv = "^1.0.1"
python-dateutil = "^2.9.0"
sbrscrape = "^0.0.10"
bs4 = "^0.0.2"
selenium-wire = "^5.1.0"
setuptools = "^75.1.0"
blinker = "1.7.0"
requests = "^2.32.5"
click = "^8.1.0"
matplotlib = "^3.9.0"
seaborn = "^0.13.0"
xgboost = ">=2.0.0,<3.0.0"
scikit-learn = ">=1.4.0,<2.0.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
